refactor(frontend): separate swap execution logic
fix: remove console.log
